\name{mod.mirt.sum}
\alias{mod.mirt.sum}
\title{IRT model (summed isomers) for estimating PFAS exposure burden}

\description{
IRT model for estimating PFAS exposure burden calculated from 7 PFAS analytes ( me-PFOSA-AcOH, PFUA, PFDeA, PFHxS, PFNA, PFOA, and PFOS) in the 2017 - 2018 NHANES. 
}
\details{
xxxxxx
}



