\name{PFAScycleJ}
\alias{PFAScycleJ}
\title{PFAS data from 2017 - 2018 NHANES}

\description{
PFAS Data obtained from the 2017 - 2018 NHANES. 
\url{https://wwwn.cdc.gov/Nchs/Data/Nhanes/Public/2017/DataFiles/PFAS_J.htm}
}
\details{

\strong{SEQN} \cr Respondent sequence number. The unique survey participant identifier in NHANES dataset.

\strong{WTSB2YR} \cr Subsample B weights. The specific sample weights for the one-third subsample of persons who had serum PFAS measurements.

\strong{me.PFOSA.AcOH} \cr 2-(N-methylperfluoroctanesulfonamido) acetic acid (Me-PFOSA-AcOH) concentration (ng/mL)

\strong{PFUA} \cr Perfluoroundecanoic acid (PFUA) concentration (ng/mL)

\strong{PFDeA} \cr Perfluorodecanoic acid (PFDeA) concentration (ng/mL)

\strong{PFHxS} \cr Perfluorohexane sulfonic acid (PFHxS) concentration (ng/mL)

\strong{PFNA} \cr Perfluorononanoic acid (PFNA) concentration (ng/mL)

\strong{n.PFOA} \cr n-perfluorooctanoic acid (n-PFOA) concentration (ng/mL)

\strong{Sm.PFOS} \cr  Perfluoromethylheptane sulfonic acid isomers (Sm-PFOS)  concentration (ng/mL)

\strong{n.PFOS} \cr n-perfluorooctane sulfonic acid (n-PFOS) concentration (ng/mL)

\strong{PFOA} \cr Summed concentration of n-PFOA and branch perfluorooctanoic acid isomers (Sb-PFOA).

\strong{PFOS} \cr Summed concentration of Sm-PFOS and n-PFOS.

\strong{Male} \cr Participant's gender (male=1, female=0).

\strong{Age} \cr Participant's age in years.

\strong{`PFAS burden`} \cr Participant's estimated PFAS burden score using summed isomers for PFOA and PFOS.

\strong{`PFAS burden (isomers)`} \cr Participant's estimated PFAS burden score using isomers for PFOA and PFOS.

}
\references{
Data documentation for 2017-2018 NHANES Perfluoroalkyl and Polyfluoroalkyl Substances (PFAS_J)
\url{https://wwwn.cdc.gov/Nchs/Data/Nhanes/Public/2017/DataFiles/PFAS_J.htm}
}



