\name{WtdQuantile}
\alias{WtdQuantile}
\title{Decile cutoffs of PFAS analytes}

\description{
Survey-weighted decile cutoffs of the PFAS analytes basing on the 2017 - 2018 NHANES cycle.
}
\details{
The survey-weighted decile cutoffs for PFAS analytes were calculated from biomonitoring data from 1915 children and adults in the 2017-2018 NHANES. (Liu et al., 2022)
}
\references{
Liu, S. H., J. R. Kuiper, Y. Chen, L. Feuerstahler, J. Teresi, and J. P. Buckley. 2022. 'Developing an Exposure Burden Score for Chemical Mixtures Using Item Response Theory, with Applications to PFAS Mixtures', Environ Health Perspect, 130: 117001. \url{https://ehp.niehs.nih.gov/doi/10.1289/EHP10125}
}



